requirements.txt
fastapi==0.110.0
uvicorn==0.30.1
google-generativeai==0.5.2
youtube-transcript-api==0.6.2
python-dotenv  # correct package name for dotenv
pytube
requests
pydantic
numpy>=1.19,<2.0  # Pin to avoid incompatibility with faiss
faiss-cpu==1.8.0
fastembed
typing-extensions  # 'typing' built-in, but typing-extensions may be needed
pydub
google-cloud-speech
google-cloud-storage
